Add Gradle,Java,Unzip to Base (Fixed build errors)

+++ imagefiles/install-gradle.sh
@@ -12,7 +12,7 @@ set -o pipefail
 
 # Download
 REV=4.3.1
-wget --progress=bar:force https://services.gradle.org/distributions/gradle-$REV-bin.zip
+wget --progress=bar:force https://services.gradle.org/distributions/gradle-$REV-bin.zip --no-check-certificate
 mkdir /opt/gradle
 unzip -d /opt/gradle gradle-$REV-bin.zip
 
@@ -20,4 +20,6 @@ unzip -d /opt/gradle gradle-$REV-bin.zip
 ln -s /opt/gradle/gradle-4.3.1/bin/gradle /bin/gradle
 
 # Clean
-rm -rf gradle-$REV-bin.zip
+if [ -f "gradle-$REV-bin.zip" ] ; then
+  rm -rf gradle-$REV-bin.zip
+fi
